Resolved issue with terrain being completely white at (0,0,0)

+++ Engine/IGraphicsAPI.h
@@ -249,7 +249,7 @@ namespace graphics
 			m_Elements.Add(obj);
 		}
 
-		size_t m_Offset = 0;
+		u32 m_Offset = 0;
 		CU::GrowingArray<InputElementDesc> m_Elements;
 
 	};
